Make the give command go through the server and target players when needed

+++ Minecraft/src/core/Scene.cpp
@@ -468,7 +468,7 @@ namespace Minecraft
 				GEventType eventType = GEventType::None;
 				while (eventType != GEventType::FrameTick && serializedEventFile && playFromEventFile)
 				{
-					int readResult = fread(buffer, sizeof(uint8), 1, serializedEventFile);
+					size_t readResult = fread(buffer, sizeof(uint8), 1, serializedEventFile);
 					if (readResult)
 					{
 						eventType = *(GEventType*)&buffer;
